requirements.txt
huggingface-hub
streamlit
langchain_community
langchain_core
python-dotenv
langchain
langserve
fastapi
uvicorn
sse_starlette